requirements.txt
# Bindings for the hermes protocol 
hermes-python>=0.2
requests
darkskylib
geopy
configparser
python-dateutil
paho-mqtt
pytz
pylint
